third_party/WebKit/LayoutTests/editing/deleting/delete_then_nbsp.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        [
            '<div contenteditable>',
            '<span>s <br><br>that| war. We  of that war.</span>',
            '</div>',
        ].join(''),
        selection => {
            for (var i = 0; i < 4; ++i)
                selection.document.execCommand('delete');
        },
        [
            '<div contenteditable>',
            '<span>s <br><br>|\u{00A0}war. We  of that war.</span>',
            '</div>',
        ].join('')),
        'Delete command converts a space to nbsp after BR.');
</script>
